native: relay finish tests in their own file (normalizer_tests.rs line budget)
